Move options cache helper into plugin options module
