+++ src/plugin/src/options/options-cache.js
